Split forest_fire.rs (575 lines) into forest_fire/ subdirectory
Split crates/simulation/src/forest_fire.rs into focused modules
under forest_fire/ for better maintainability:
- mod.rs: thin facade with re-exports
- resources.rs: ForestFireGrid, ForestFireStats
- constants.rs: tuning constants and fire_hash helper
- helpers.rs: neighbors4, neighbors8, is_near_industrial
- systems.rs: update_forest_fire system and ForestFirePlugin
- tests.rs: all unit tests
