+++ crates/simulation/src/forest_fire/mod.rs
@@ -1,0 +1,8 @@
+mod constants;
+mod helpers;
+mod resources;
+mod systems;
+mod tests;
+
+pub use resources::{ForestFireGrid, ForestFireStats};
+pub use systems::{update_forest_fire, ForestFirePlugin};
