+++ crates/simulation/src/forest_fire/systems.rs
@@ -2,105 +2,16 @@ use bevy::prelude::*;
 
 use crate::config::{GRID_HEIGHT, GRID_WIDTH};
 use crate::fire::FireGrid;
-use crate::grid::{CellType, WorldGrid, ZoneType};
+use crate::grid::{CellType, WorldGrid};
 use crate::land_value::LandValueGrid;
 use crate::trees::TreeGrid;
 use crate::weather::{Weather, WeatherCondition};
 use crate::wind::WindState;
 use crate::TickCounter;
 
-// =============================================================================
-// Resources
-// =============================================================================
-
-/// Per-cell forest fire intensity grid. 0 = no fire, 1-255 = burning intensity.
-#[derive(Resource)]
-pub struct ForestFireGrid {
-    pub intensities: Vec<u8>,
-    pub width: usize,
-    pub height: usize,
-}
-
-impl Default for ForestFireGrid {
-    fn default() -> Self {
-        Self {
-            intensities: vec![0; GRID_WIDTH * GRID_HEIGHT],
-            width: GRID_WIDTH,
-            height: GRID_HEIGHT,
-        }
-    }
-}
-
-impl ForestFireGrid {
-    #[inline]
-    pub fn get(&self, x: usize, y: usize) -> u8 {
-        self.intensities[y * self.width + x]
-    }
-
-    #[inline]
-    pub fn set(&mut self, x: usize, y: usize, val: u8) {
-        self.intensities[y * self.width + x] = val;
-    }
-}
-
-/// Tracks forest fire statistics over time.
-#[derive(Resource, Default, Debug)]
-pub struct ForestFireStats {
-    pub active_fires: u32,
-    pub total_area_burned: u64,
-    pub fires_this_month: u32,
-}
-
-// =============================================================================
-// Constants
-// =============================================================================
-
-/// How often (in ticks) the forest fire system runs.
-const FIRE_UPDATE_INTERVAL: u64 = 10;
-
-/// Lightning strike chance per cell during storms (per update cycle).
-/// Applied only to cells with trees. Very low: checked against hash % 100_000.
-const LIGHTNING_CHANCE_PER_CELL: u64 = 2; // 2 in 100_000
-
-/// Chance for fire to spread from a burning building (FireGrid) to adjacent forest cells.
-/// Checked against hash % 1000.
-const BUILDING_FIRE_SPREAD_THRESHOLD: u64 = 50; // 5% per neighbor
-
-/// Chance for spontaneous ignition near industrial zones in hot weather.
-/// Checked against hash % 100_000.
-const INDUSTRIAL_IGNITION_THRESHOLD: u64 = 5; // 5 in 100_000
-
-/// Base fire intensity when a cell first ignites.
-const INITIAL_INTENSITY: u8 = 30;
-
-/// How much intensity decreases per update tick (natural burnout).
-const BURNOUT_RATE: u8 = 2;
-
-/// How much rain reduces fire intensity per update tick.
-const RAIN_REDUCTION: u8 = 8;
-
-/// How much storm reduces fire intensity per update tick.
-const STORM_REDUCTION: u8 = 15;
-
-/// Threshold above which a forest fire can ignite a building.
-const BUILDING_IGNITION_THRESHOLD: u8 = 100;
-
-/// Land value penalty applied per burning cell in radius 3.
-const LAND_VALUE_PENALTY: u8 = 5;
-
-// =============================================================================
-// Deterministic pseudo-random helper
-// =============================================================================
-
-/// Deterministic hash for a given tick and cell index.
-/// Returns a value in [0, modulus).
-#[inline]
-fn fire_hash(tick: u64, cell_index: usize, salt: u64) -> u64 {
-    tick.wrapping_mul(7919)
-        .wrapping_add(cell_index as u64)
-        .wrapping_add(salt)
-        .wrapping_mul(2654435761)
-}
+use super::constants::*;
+use super::helpers::{is_near_industrial, neighbors4, neighbors8};
+use super::resources::{ForestFireGrid, ForestFireStats};
 
 // =============================================================================
 // Systems
@@ -343,221 +254,8 @@ pub fn update_forest_fire(
 }
 
 // =============================================================================
-// Helpers
+// Plugin
 // =============================================================================
-
-/// Returns the valid 4-connected neighbors of cell (x, y).
-fn neighbors4(x: usize, y: usize) -> Vec<(usize, usize)> {
-    let mut result = Vec::with_capacity(4);
-    if x > 0 {
-        result.push((x - 1, y));
-    }
-    if x + 1 < GRID_WIDTH {
-        result.push((x + 1, y));
-    }
-    if y > 0 {
-        result.push((x, y - 1));
-    }
-    if y + 1 < GRID_HEIGHT {
-        result.push((x, y + 1));
-    }
-    result
-}
-
-/// Returns the valid 8-connected neighbors of cell (x, y).
-fn neighbors8(x: usize, y: usize) -> Vec<(usize, usize)> {
-    let mut result = Vec::with_capacity(8);
-    for dy in -1i32..=1 {
-        for dx in -1i32..=1 {
-            if dx == 0 && dy == 0 {
-                continue;
-            }
-            let nx = x as i32 + dx;
-            let ny = y as i32 + dy;
-            if nx >= 0 && ny >= 0 && (nx as usize) < GRID_WIDTH && (ny as usize) < GRID_HEIGHT {
-                result.push((nx as usize, ny as usize));
-            }
-        }
-    }
-    result
-}
-
-/// Checks if there is an industrial zone within `radius` cells of (x, y).
-fn is_near_industrial(grid: &WorldGrid, x: usize, y: usize, radius: i32) -> bool {
-    for dy in -radius..=radius {
-        for dx in -radius..=radius {
-            if dx.abs() + dy.abs() > radius {
-                continue;
-            }
-            let nx = x as i32 + dx;
-            let ny = y as i32 + dy;
-            if nx >= 0
-                && ny >= 0
-                && (nx as usize) < GRID_WIDTH
-                && (ny as usize) < GRID_HEIGHT
-                && grid.get(nx as usize, ny as usize).zone == ZoneType::Industrial
-            {
-                return true;
-            }
-        }
-    }
-    false
-}
-
-// =============================================================================
-// Tests
-// =============================================================================
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_forest_fire_grid_default() {
-        let grid = ForestFireGrid::default();
-        assert_eq!(grid.width, GRID_WIDTH);
-        assert_eq!(grid.height, GRID_HEIGHT);
-        assert_eq!(grid.intensities.len(), GRID_WIDTH * GRID_HEIGHT);
-        assert!(grid.intensities.iter().all(|&v| v == 0));
-    }
-
-    #[test]
-    fn test_forest_fire_grid_get_set() {
-        let mut grid = ForestFireGrid::default();
-        assert_eq!(grid.get(10, 20), 0);
-        grid.set(10, 20, 150);
-        assert_eq!(grid.get(10, 20), 150);
-    }
-
-    #[test]
-    fn test_forest_fire_grid_boundary() {
-        let mut grid = ForestFireGrid::default();
-        grid.set(0, 0, 255);
-        assert_eq!(grid.get(0, 0), 255);
-        grid.set(GRID_WIDTH - 1, GRID_HEIGHT - 1, 100);
-        assert_eq!(grid.get(GRID_WIDTH - 1, GRID_HEIGHT - 1), 100);
-    }
-
-    #[test]
-    fn test_forest_fire_stats_default() {
-        let stats = ForestFireStats::default();
-        assert_eq!(stats.active_fires, 0);
-        assert_eq!(stats.total_area_burned, 0);
-        assert_eq!(stats.fires_this_month, 0);
-    }
-
-    #[test]
-    fn test_fire_hash_deterministic() {
-        let a = fire_hash(100, 5000, 0);
-        let b = fire_hash(100, 5000, 0);
-        assert_eq!(a, b);
-    }
-
-    #[test]
-    fn test_fire_hash_varies_with_inputs() {
-        let a = fire_hash(100, 5000, 0);
-        let b = fire_hash(101, 5000, 0);
-        let c = fire_hash(100, 5001, 0);
-        let d = fire_hash(100, 5000, 1);
-        // All should be different (extremely high probability)
-        assert_ne!(a, b);
-        assert_ne!(a, c);
-        assert_ne!(a, d);
-    }
-
-    #[test]
-    fn test_neighbors4_center() {
-        let n = neighbors4(128, 128);
-        assert_eq!(n.len(), 4);
-        assert!(n.contains(&(127, 128)));
-        assert!(n.contains(&(129, 128)));
-        assert!(n.contains(&(128, 127)));
-        assert!(n.contains(&(128, 129)));
-    }
-
-    #[test]
-    fn test_neighbors4_corner() {
-        let n = neighbors4(0, 0);
-        assert_eq!(n.len(), 2);
-        assert!(n.contains(&(1, 0)));
-        assert!(n.contains(&(0, 1)));
-    }
-
-    #[test]
-    fn test_neighbors8_center() {
-        let n = neighbors8(128, 128);
-        assert_eq!(n.len(), 8);
-        // Check all 8 directions
-        assert!(n.contains(&(127, 127)));
-        assert!(n.contains(&(128, 127)));
-        assert!(n.contains(&(129, 127)));
-        assert!(n.contains(&(127, 128)));
-        assert!(n.contains(&(129, 128)));
-        assert!(n.contains(&(127, 129)));
-        assert!(n.contains(&(128, 129)));
-        assert!(n.contains(&(129, 129)));
-    }
-
-    #[test]
-    fn test_neighbors8_corner() {
-        let n = neighbors8(0, 0);
-        assert_eq!(n.len(), 3);
-        assert!(n.contains(&(1, 0)));
-        assert!(n.contains(&(0, 1)));
-        assert!(n.contains(&(1, 1)));
-    }
-
-    #[test]
-    fn test_constants_valid() {
-        assert!(FIRE_UPDATE_INTERVAL > 0);
-        assert!(INITIAL_INTENSITY > 0);
-        assert!(BURNOUT_RATE > 0);
-        assert!(RAIN_REDUCTION > BURNOUT_RATE);
-        assert!(STORM_REDUCTION > RAIN_REDUCTION);
-        assert!(BUILDING_IGNITION_THRESHOLD > INITIAL_INTENSITY);
-    }
-
-    #[test]
-    fn test_is_near_industrial() {
-        use crate::grid::WorldGrid;
-        let mut grid = WorldGrid::new(GRID_WIDTH, GRID_HEIGHT);
-        // No industrial zones by default
-        assert!(!is_near_industrial(&grid, 128, 128, 3));
-
-        // Place an industrial zone
-        grid.get_mut(130, 128).zone = ZoneType::Industrial;
-        assert!(is_near_industrial(&grid, 128, 128, 3));
-        assert!(!is_near_industrial(&grid, 128, 128, 1));
-    }
-
-    #[test]
-    fn test_burnout_reduces_intensity() {
-        // Simulate burnout: intensity should decrease by BURNOUT_RATE
-        let intensity: u8 = 50;
-        let after_burnout = intensity.saturating_sub(BURNOUT_RATE);
-        assert_eq!(after_burnout, 50 - BURNOUT_RATE);
-    }
-
-    #[test]
-    fn test_rain_extinguishes_small_fires() {
-        // A small fire (intensity = 5) should be extinguished by rain
-        let intensity: u8 = 5;
-        let after = intensity
-            .saturating_sub(BURNOUT_RATE)
-            .saturating_sub(RAIN_REDUCTION);
-        assert_eq!(after, 0);
-    }
-
-    #[test]
-    fn test_storm_extinguishes_moderate_fires() {
-        // A moderate fire (intensity = 15) should be extinguished by storm
-        let intensity: u8 = 15;
-        let after = intensity
-            .saturating_sub(BURNOUT_RATE)
-            .saturating_sub(STORM_REDUCTION);
-        assert_eq!(after, 0);
-    }
-}
 
 pub struct ForestFirePlugin;
